requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.7.0
django-currentuser==0.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
